Apply optimizations for a single file
